requirements.txt
python-fasthtml
pandas
openpyxl

